requirements.txt
Flask~=2.1.0
Flask-MySQL~=1.5.2
SQLAlchemy~=1.3.18
gunicorn~=20.0.4
pyOpenSSL~=19.1.0
colormath~=3.0.0
Pillow~=9.3.0
scikit-learn~=1.2.0
